Cleanup some
